Initial GraphQL dependency removal
